% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{AS.models}
\alias{AS.models}
\title{Models for AS endpoints}
\usage{
AS.models(
  biodb,
  train.control = NULL,
  target = c("ProgressedToTreatment", "BiopsyUpgraded", "Prostatectomy"),
  metric = c("F", "Accuracy", "PR-AUC", "Kappa", "Precision", "Recall", "ROC-AUC",
    "Sens", "Spec"),
  models = c("xgb", "gbm", "rpart"),
  exclude.vars = NULL,
  include.vars = NULL,
  predict.missing = FALSE,
  seed = NULL,
  rpart.cost = NULL,
  rm.NoUpgradeAndProgressed = FALSE,
  reduced.model = FALSE,
  baseline.model = FALSE,
  suffix = ""
)
}
\arguments{
\item{include.vars}{Include additional variables for the models}

\item{reduced.model}{Use the reduced set of features for the model}

\item{baseline.model}{Use the clinically relevant demographic features for the model}
}
\value{

}
\description{
Models for AS endpoints
}
