% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{data}
\name{AS.train.control}
\alias{AS.train.control}
\title{Caret train control settings}
\format{
An object of class \code{list} of length 27.
}
\usage{
AS.train.control
}
\description{
Caret train control settings
}
\keyword{datasets}
