% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{camel.to.spaces}
\alias{camel.to.spaces}
\title{Convert camel case to string with spaces. Note does not work with multiple capitals
Idea from: https://stackoverflow.com/a/8407047/1351718}
\usage{
camel.to.spaces(x, replace = " ")
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Convert camel case to string with spaces. Note does not work with multiple capitals
Idea from: https://stackoverflow.com/a/8407047/1351718
}
\examples{
camel.to.spaces(c('BiopsyUpgraded', 'ProgressedToTreatment'))
}
