% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.models.R
\name{compare.var.imp}
\alias{compare.var.imp}
\title{Compare variable importance across multiple models}
\usage{
compare.var.imp(x, include.ranks = FALSE)
}
\arguments{
\item{ranks}{}
}
\value{

}
\description{
Compare variable importance across multiple models
}
