% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{create.heatmap.AS}
\alias{create.heatmap.AS}
\title{Creates a correlation heatmap for the AS cohort}
\usage{
create.heatmap.AS(biodb, ...)
}
\arguments{
\item{biodb}{the data frame of the patients}

\item{...}{Additional arguments to pass to \code{BoutrosLab.plotting.general::create.heatmap}}
}
\value{
Output from \code{BoutrosLab.plotting.general::create.heatmap}
}
\description{
Creates a correlation heatmap for the AS cohort
}
\examples{
biodb <- default.load.data(onlyBiodb = TRUE)
create.heatmap.AS(biodb)
create.heatmap.AS(biodb,
  filename = here('figures/corr_heatmap.tiff'))
}
