% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{custom.summary}
\alias{custom.summary}
\title{Custom summary for caret model}
\usage{
custom.summary(data, lev = NULL, model = NULL)
}
\arguments{
\item{model}{}
}
\value{

}
\description{
Custom summary for caret model
}
