% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.data.R
\name{default.load.data}
\alias{default.load.data}
\title{Loads the data for analysis with default file names}
\usage{
default.load.data(onlyBiodb = FALSE)
}
\arguments{
\item{onlyBiodb}{boolean indicating if we should only return the biodb (main data)}
}
\description{
Loads the data for analysis with default file names
}
