% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{label.or.name}
\alias{label.or.name}
\title{Get the label or the name of a variable or data frame. If x is a data frame, then return the label or column name.}
\usage{
label.or.name(x)
}
\arguments{
\item{x}{the variable or data frame we want to get either the label or name from.}
}
\description{
Get the label or the name of a variable or data frame. If x is a data frame, then return the label or column name.
}
\examples{
x <- data.frame(x = 1:4, y = 5:8)
attr(x$x, 'label') <- 'Label 1'
label.or.name(x$x)
#> "Label 1"
label.or.name(x$y)
#> x$y
label.or.name(x)
#> "Label 1" "y"
}
