% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.data.R
\name{load.biomarker.categories}
\alias{load.biomarker.categories}
\title{Read in the biomarker categories}
\usage{
load.biomarker.categories(file.name = "biomarkers_categories.xlsx")
}
\arguments{
\item{file.name}{}
}
\value{

}
\description{
Read in the biomarker categories
}
