% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.data.R
\name{load.data.AS}
\alias{load.data.AS}
\title{Loads the data for the Active Surveillance project}
\usage{
load.data.AS(
  biomark.path = NULL,
  biomark.key.path = NULL,
  genetics.path = NULL
)
}
\arguments{
\item{biomark.key.path}{path to the Biomarkers key xlsx file}

\item{biomark.pathBiomarkers}{xlsx file}

\item{genetics}{path to the genetics xlsx file}
}
\description{
Loads the data for the Active Surveillance project
}
