% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.models.R
\name{optimal.threshold.train}
\alias{optimal.threshold.train}
\title{Computes the optimal threshold based on a given cost matrix.}
\usage{
optimal.threshold.train(
  model,
  cost.matrix = matrix(c(0, 1, 1, 0), byrow = TRUE, nrow = 2),
  thresholds = c(0.25, 0.5, 0.75),
  smooth = TRUE
)
}
\arguments{
\item{smooth}{}
}
\value{

}
\description{
Computes the optimal threshold based on a given cost matrix.
}
