% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{str.match}
\alias{str.match}
\title{Returns the group matches from a regular expression on a vector}
\usage{
\method{str}{match}(x, pattern)
}
\arguments{
\item{x}{the vector we want to match on}

\item{pattern}{regular expression with groups}
}
\description{
Returns the group matches from a regular expression on a vector
}
