% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.models.R
\name{summarize.seq.models}
\alias{summarize.seq.models}
\title{Summarize the sequence models}
\usage{
summarize.seq.models(models, models.roc = NULL)
}
\arguments{
\item{models.roc}{}
}
\value{

}
\description{
Summarize the sequence models
}
