% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tte.R
\name{surv.format}
\alias{surv.format}
\title{Converts event times and end of study times into survival::Surv format}
\usage{
surv.format(event.times, end.times)
}
\arguments{
\item{event.times}{The times of the events. Expects NA if event did not occur.}

\item{end.times}{The last follow up time.}
}
\value{
a list with keys \code{time} and \code{event}
}
\description{
Converts event times and end of study times into survival::Surv format
}
\examples{
x <- c(NA, 1, NA, 2)
y <- c(3, 4, 4.5, 7)
do.call(survival::Surv, surv.format(x, y))
}
