% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.models.R
\name{threshold.summary.stats}
\alias{threshold.summary.stats}
\title{Compute the threshold summary stats per fold}
\usage{
threshold.summary.stats(model, threshold, per.fold = FALSE)
}
\arguments{
\item{threshold}{}
}
\value{

}
\description{
Compute the threshold summary stats per fold
}
