% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.models.R
\name{var.imp.combine}
\alias{var.imp.combine}
\title{Combine the variable importance scores from multiple models into one data frame}
\usage{
var.imp.combine(x, order = FALSE, rev.cols = FALSE)
}
\arguments{
\item{x}{list of caret train classes}
}
\value{

}
\description{
Combine the variable importance scores from multiple models into one data frame
}
