% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.models.R
\name{var.imp}
\alias{var.imp}
\title{Compute the variable importance for GBM, XGB, rpart}
\usage{
var.imp(x, onlyImportance = TRUE, squash = TRUE)
}
\arguments{
\item{onlyImportance}{boolean indicating whether only a data frame with the importance scores be returned? FALSE returns class 'caret::varImp.train'.}

\item{squash}{Should multiple dummy variables be squashed into one variable? (Relevant for xgb models)}
}
\value{

}
\description{
Compute the variable importance for GBM, XGB, rpart
}
