% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.data.R
\name{load.data.AS}
\alias{load.data.AS}
\title{Loads the data for the Active Surveillance project}
\usage{
load.data.AS(
  biomark.path,
  biomark.key.path,
  genetics.path,
  biomark.categories.path,
  onlyBiodb = FALSE
)
}
\arguments{
\item{biomark.key.path}{path to the Biomarkers key xlsx file}

\item{onlyBiodb}{boolean indicating if we should only return the biodb (main data)}

\item{biomark.pathBiomarkers}{xlsx file}

\item{genetics}{path to the genetics xlsx file}
}
\description{
Loads the data for the Active Surveillance project
}
